requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\hp\Documents\Projects\Prime-Test-Complexity-Analysis\prime_counting_complexity_analysis.py: 6
# C:\Users\hp\Documents\Projects\Prime-Test-Complexity-Analysis\prime_test_complexity_analysis.py: 6
numpy == 1.20.3

# C:\Users\hp\Documents\Projects\Prime-Test-Complexity-Analysis\prime_counting_complexity_analysis.py: 3
# C:\Users\hp\Documents\Projects\Prime-Test-Complexity-Analysis\prime_test_complexity_analysis.py: 3
pandas_stubs == 1.1.0.12
